+++ CHANGELOG.md
@@ -6,10 +6,15 @@
 * [📜 Changelog](#-changelog)
 * [📖 Table of contents](#-table-of-contents)
 * [0.x.y](#0xy)
+  * [0.0.2](#002)
   * [0.0.1](#001)
 <!-- TOC -->
 
 # 0.x.y
+
+## 0.0.2
+- Zod based validation system
+- ZodError to ProblemDetail system
 
 ## 0.0.1
 - Initial release.